requirements.txt
numpy
sounddevice
mlx-whisper
webrtcvad
python-dotenv
sentence-transformers
streamlit
